pyproject.toml
[project]
name = "webinterviewrecorder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.121.2",
    "firebase-admin>=7.1.0",
    "pydantic>=2.12.4",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "unidecode>=1.4.0",
    "uvicorn>=0.38.0",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.2",
]
